refactor(app): Restrict FastHTML imports in routes
Imporing everything FastHTML in routes strikes me as
inefficient when everything tends to happen in the
page or the model anyway, and the only reason we need
the FastHTML object is to create more Starlette objects.

This commit budges page creation to pages/, allowing us
to have the only imported FastHTML object to be the
FastHTML routes handler.