feature(app): Add developer Telegram alerting
This commit adds a simple call to the Telegram Bot
API webook. It is intended to alert a developer of
issues on the prod service.

To set it up, you will need a bot API key from
Telegram, see the bot API <https://core.telegram.org/bots>,
and a chat ID to send messages to.

If either of these things aren't specified, the alert isn't sent.

As an example, it has been added to the 500 error exception
handler. The try/except is important here otherwise if you have no
Telegram bot config set up, the user won't see the pretty 500 error
page.