feature(auth): Replace password auth with GitHub OAuth
This big bang commit removes the password authentication
system and replaces it with an OAuth authentication
system using GitHub's OAuth service.

As part of this, the following has been removed:

- The registration and login email/password
form;

- E-Mail authentication helpers.

The migration does the following:

- Copies the users table prior to migration to
an archive table;

- Adds GitHub-specific user profile information
to the users table;

- Drops password and salt columns from user;

- Truncates the users table.

Also included here are changes to make the logbook
functionality work, using GitHub profile information
instead of an email address, along with unit
test updates and documenation updates.